work in progress update

+++ src/main/java/com/lip6/services/IServiceContact.java
@@ -3,6 +3,7 @@ package com.lip6.services;
 import java.util.ArrayList;
 import java.util.List;
 
+import com.lip6.entities.Address;
 import com.lip6.entities.Contact;
 import com.lip6.entities.ContactGroup;
 import com.lip6.entities.PhoneNumber;
@@ -15,10 +16,9 @@ public interface IServiceContact {
 	public boolean editContact(Contact contact);
 	public boolean updateContact(Contact contact);
 	public boolean deleteContact(int id);
-	public ArrayList<PhoneNumber> getPhonesByIdContact(int idContact);
-	public ArrayList<ContactGroup> getGroupesByIdContact(int idContact);
+	public ArrayList<PhoneNumber> getPhonesByIdContact(Long idContact);
+	public ArrayList<ContactGroup> getGroupesByIdContact(Long idContact);
 	public ArrayList<PhoneNumber> getPhones();
 	public ArrayList<ContactGroup> getGroupes();
-
 
 }
